# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Green Shop Api Documentaion By Swagger",
        "description": "Green Shop Service Api Documentation (BehAndam Admin Services)",
        "contact": {
            "email": "saber.tabatabaee@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://debug.behaminplus.ir/shop-service",
            "description": "debug behaminplus API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/shop-service",
            "description": "develop behaminplus API Server"
        },
        {
            "url": "https://behaminplus.ir/shop-service",
            "description": "stage behaminplus API Server"
        },
        {
            "url": "https://behandam.kermany.com/shop-service",
            "description": "PRODUCTION behaminplus API Server"
        },
        {
            "url": "https://shop.pardisania.ir",
            "description": "pardisania API Server"
        },
        {
            "url": "https://beitak.ir",
            "description": "beitak API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "LOCAL API Server"
        }
    ],
    "paths": {
...
        "/api/products/{productId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Returns product by id as json",
                "description": "Returns products by id",
                "operationId": "getProductsbyid",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/users/{userId}/products/payed": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Returns payed products",
                "description": "Returns payed products",
                "operationId": "getPayedProducts",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "userId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Use username & password for token",
                "name": "BearerAuth",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}